+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Added support of the `qs2` format (#865, @atsyplenkov).
 
+* Added download progress bar for `board_url()` (#872, @lbm364dl).
+
 # pins 1.4.1
 
 * Support new `preview_data` parameter for pin previews on Posit Connect (#850).